+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # mParticle iOS SDK CHANGELOG
 
+## 5.3.1
+
+* Determining whether to forward an app delegate call to the old deep-linking method
+* Forwarding event attributes as user attributes to Appboy
+
 ## 5.3.0
 
 * Support for [AppsFlyer](https://www.appsflyer.com) as a kit